Implemented Year.java and changed so that the methods in Year and Month throw IllegalArgumentExceptions when called on with invalid dates or weights. Haven't tested either yet, but will probably do that next.

+++ application/YearADT.java
@@ -54,7 +54,7 @@ public interface YearADT {
 	 * Returns the average monthly weight for this year
 	 * @return average monthly weight in this year
 	 */
-	public int avgMonthWeight();
+	public float avgMonthWeight();
 	
 	/**
 	 * Returns the total amount of weight stored from the given starting date
